+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - UIColor from HEX
 - Keyboard Layout Guide
 - Cron Jobs
+- Location Services: Convenience wrapper for authorization 
 
 ### Changed
 - Return cached metrics